Update protobuf dep in bazel build to v30.1 (#521)
Fixes downstream build issues with projects using protobuf v30.1. Some of the return types in protobuf were changed to std::string_view instead of std::string in v30. Some usages in core/generator_lite/Generator.cc need to be updated as a result.

I also had to update the version of the bazel-contrib script we use for bazel CI to circumvent an error on bazel 8.0.0rc1 (brought in transitively with the protobuf upgrade). With this change, bazel 8.x CI now uses bazel 8.0.0.

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>